refactor: Refactors optimizations-backend
Details:
- Removes alpha code
- Cleans up unnecessary modules
