+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization-backend/src/index.ts
@@ -13,5 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './service/router';
 export { resourceOptimizationPlugin as default } from './plugin';
